+++ Lib/fontTools/varLib/instancer/__main__.py
@@ -1,0 +1,5 @@
+import sys
+from fontTools.varLib.instancer import main
+
+if __name__ == "__main__":
+    sys.exit(main())
