+++ Tests/varLib/instancer/conftest.py
@@ -1,0 +1,13 @@
+import os
+from fontTools import ttLib
+import pytest
+
+
+TESTDATA = os.path.join(os.path.dirname(__file__), "data")
+
+
+@pytest.fixture
+def varfont():
+    f = ttLib.TTFont()
+    f.importXML(os.path.join(TESTDATA, "PartialInstancerTest-VF.ttx"))
+    return f
